[03665] Extract GetDashboardData to DashboardRepository
- Created DashboardRepository class in new Repositories namespace
- Moved 148-line GetDashboardData method from PlanDatabaseService
- PlanDatabaseService now delegates to DashboardRepository
- No logic changes, pure refactoring for separation of concerns
